At #444 Separate Code Maximum Text Size Calculation Logic for Reusability in Dygraphs Legends
